Create app.json in the root directory
Using heroku 'review apps' requires this file to exist in the repository,
so we cannot just produce it in the dist/ directory. Instead, produce it
in the root, and check it into version control.

Git compares file content, and so as long as nothing changes we should be
fine -- and when something changes we should notice that app.json also changes.