Change the default access log format to 'common'
The 'short' format doesn't contain timestamps, which are extremely important in production
to be able to correlate requests with other events.

Note that it is expected for developers to use the 'dev' format via their startup systems.

Also note that we do _not_ use the 'combined' log format, as the referer may contain information
we do not want to have in our logs for privacy reasons.